refactor(core): move curated API types to @3amoncall/core
Move all curated API types (runtime-map, extended-incident, evidence)
from console-local definitions to packages/core/src/schemas/curated-api.ts.
Console's curated-types.ts is now a pure re-export from @3amoncall/core.

This establishes a single source of truth for the curated API contract
shared by receiver, diagnosis, and frontend.
