+++ packages/core/src/index.ts
@@ -4,3 +4,4 @@ export * from "./schemas/diagnosis-result.js";
 export * from "./schemas/thin-event.js";
 export * from "./schemas/extracted-span.js";
 export * from "./schemas/anomalous-signal.js";
+export * from "./schemas/curated-api.js";
